rename 'main' file
